README improvements.
Example improvement.
cmake improvement.
github workflow tuning.

+++ Examples/ARM/arm_class_marks_example/arm_class_marks_example_f32.c
@@ -212,7 +212,7 @@ int32_t main()
   arm_var_f32(testOutput, numStudents, &var);
 
 #if defined(SEMIHOSTING)
-  printf("mean = %f, std = %f\n",mean,std);
+  printf("mean = %f, std = %f\n",(double)mean,(double)std);
 #endif
 
 #if !defined(SEMIHOSTING)
