+++ Examples/ARM/arm_convolution_example/math_helper.c
@@ -100,7 +100,7 @@ float arm_snr_f32(float *pRef, float *pTest, uint32_t buffSize)
     }
 
 
-  SNR = 10 * log10 (EnergySignal / EnergyError);
+  SNR = 10 * (float32_t)log10 ((double)EnergySignal / (double)EnergyError);
 
   return (SNR);
 
@@ -167,7 +167,7 @@ uint32_t arm_compare_fixed_q15(q15_t *pIn, q15_t *pOut, uint32_t numSamples)
 {
   uint32_t i;
   int32_t diff, diffCrnt = 0;
-  uint32_t maxDiff = 0;
+  int32_t maxDiff = 0;
 
   for (i = 0; i < numSamples; i++)
   {
@@ -195,7 +195,7 @@ uint32_t arm_compare_fixed_q31(q31_t *pIn, q31_t * pOut, uint32_t numSamples)
 {
   uint32_t i;
   int32_t diff, diffCrnt = 0;
-  uint32_t maxDiff = 0;
+  int32_t maxDiff = 0;
 
   for (i = 0; i < numSamples; i++)
   {
